requirements.txt
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.0.2
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.2.0
django-ranged-response==0.2.0
django-reversion==3.0.3
django-simple-captcha==0.5.10
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.9.2
jdcal==1.4
odfpy==1.4.0
openpyxl==2.6.0
Pillow==5.4.1
PyMySQL==0.9.3
pytz==2018.9
PyYAML==3.13
six==1.12.0
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.2.0
xlwt==1.3.0



